requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic[email]==2.12.5
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
supabase==2.28.0
python-dotenv==1.2.2
playwright==1.58.0
apscheduler==3.11.2
httpx==0.28.1
beautifulsoup4==4.14.3
lxml==6.0.2
openai==2.26.0
groq==1.1.1
